Allow caching of individual records

+++ registry/src/orca/_functions/orca_export_functions.php
@@ -3186,9 +3186,9 @@ function getSearchBaseScore($registry_object_key)
 	$baseScore = eBOOST_DEFAULT_BASE;
 
 	// number of related objects:
-	$number_of_related_objects = getRelatedObjectCount($registry_object_key);
-	$number_of_related_objects -= getMinorImpactRelatedObjectCount($registry_object_key); //remove close relationships ("hasPart")
-	$baseScore += eBOOST_RELATED_OBJECT_ADJUSTMENT * (int) $number_of_related_objects;
+	$number_of_related_objects = (int) getRelatedObjectCount($registry_object_key);
+	$number_of_related_objects -= (int) getMinorImpactRelatedObjectCount($registry_object_key); //remove close relationships ("hasPart")
+	$baseScore += eBOOST_RELATED_OBJECT_ADJUSTMENT * $number_of_related_objects;
 
 	// number of INCOMING related objects (objects that relate to us):
 	$number_of_related_objects = getIncomingRelatedObjectCount($registry_object_key);
